registration updates and organization of files
