--- account_forms.py
@@ -1,21 +1,0 @@
-"""Forms for account creation and management"""
-
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField, BooleanField
-from wtforms.validators import DataRequired, Length, Email, EqualTo
-
-
-class RegistrationForm(FlaskForm):
-    """Form for registering an account"""
-    username = StringField('Username', validators=[DataRequired(), Length(min = 2, max = 20)])
-    email = StringField('Email Address', validators=[DataRequired(), Email()])
-    password = PasswordField('Password', validators=[DataRequired()])
-    repeatPassword = PasswordField('Confirm Password', validators=[DataRequired(), EqualTo('password')])
-    submit = SubmitField('Create Account')
-
-class LoginForm(FlaskForm):
-    #Form to Login
-    username = StringField('Username', validators=[DataRequired(), Length(min = 2, max = 20)])
-    password = PasswordField('Password', validators=[DataRequired()])
-    remember = BooleanField('Remember Me')
-    submit = SubmitField('Create Account')
