--- login.html
@@ -1,71 +1,0 @@
-{% extends "index.html" %}
-{% block content%}
-
-<div class="content_section">
-    <form method="post" action="">
-        {{ form.hidden_tag() }}
-        <fieldset class="form-group">
-            <legend class="border-bottom mb-4"> Log In </legend>
-
-            <div class="form-group">
-                <!-- Login using Username, throws error if invalid-->
-                {{form.username.label(class = "form-control-label")}}
-
-                {% if form.username.errors %}
-                {{ form.username(class="form-control form-control-lg is-invalid")}}
-                <div class="invalid-feedback">
-                    {% for error in form.username.errors %}
-                    <span>
-                        {{error}}
-                    </span>
-                    {% endfor %}
-                </div>
-                {% else %}
-                {{form.username(class = "form-control form-control-lg")}}
-                {% endif %}
-            </div>
-
-            <div class="form-group">
-                <!-- Gets Password, throws error if invalid-->
-                {{form.password.label(class = "form-control-label")}}
-
-                {% if form.password.errors %}
-                {{ form.password(class="form-control form-control-lg is-invalid")}}
-                <div class="invalid-feedback">
-                    {% for error in form.password.errors %}
-                    <span>
-                        {{error}}
-                    </span>
-                    {% endfor %}
-                </div>
-                {% else %}
-                {{form.password(class = "form-control form-control-lg")}}
-                {% endif %}
-            </div>
-
-            <!-- I'm not sure if we need to implement this, get rid of if not needed-->
-            <div class="form-check">
-                <!-- If user wants to be remembered the next time they log in-->
-                {{ form.remember(class="form-check-input")}}
-                {{ form.remember.label(class="form-check-label")}}
-            </div>
-            <!-- End of Possible Extranious Code-->
-        </fieldset>
-
-        <div class="form-group">
-            <!-- Button to Submit info to log in-->
-            {{ form.submit(class="btn btn-outline-info") }}
-        </div>
-
-        <!-- I'm not sure if we need to implement this, get rid of if not needed-->
-        <small class="text-muted ml-2">
-            <a href="#">
-                Forgot Password?
-            </a>
-        </small>
-        <!-- End of Possible Extranious Code-->
-
-    </form>
-</div>
-
-{% endblock content %}
